# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecolibary API",
        "description": " API pour gérer une collection de livres classés par catégories. Le système permet de suivre la disponibilité des ouvrages et de surveiller l'état physique de la collection en suivant précisément les exemplaires dégradés.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Liste de tous les livres",
                "description": "Récupère la liste complète de tous les livres disponibles dans la bibliothèque, avec leur catégorie.",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "⚠️ Admin Only - Ajouter un nouveau livre",
                "description": "Permet à un administrateur d'ajouter un nouveau livre dans la bibliothèque.",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "category_id",
                                    "total_copies",
                                    "available_copies",
                                    "degraded_copies"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Le Seigneur des Anneaux"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "J.R.R. Tolkien"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une épopée fantastique se déroulant en Terre du Milieu.",
                                        "nullable": true
                                    },
                                    "total_copies": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "available_copies": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "degraded_copies": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "403": {
                        "description": "Accès refusé - Réservé aux administrateurs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Afficher les détails d'un livre",
                "description": "Récupère les informations détaillées d'un livre spécifique via son identifiant.",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique du livre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "⚠️ Admin Only - Mettre à jour un livre",
                "description": "Permet à un administrateur de modifier les informations d'un livre existant.",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique du livre à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Les Misérables - Édition Révisée"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Victor Hugo"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un chef-d'œuvre de la littérature française.",
                                        "nullable": true
                                    },
                                    "total_copies": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "available_copies": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "degraded_copies": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "403": {
                        "description": "Accès refusé - Réservé aux administrateurs"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "⚠️ Admin Only - Supprimer un livre",
                "description": "Permet à un administrateur de supprimer définitivement un livre de la bibliothèque.",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique du livre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "403": {
                        "description": "Accès refusé - Réservé aux administrateurs"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "description": "Modèle représentant un livre dans la bibliothèque EcoLibrary",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Le Petit Prince"
                    },
                    "author": {
                        "type": "string",
                        "example": "Antoine de Saint-Exupéry"
                    },
                    "slug": {
                        "type": "string",
                        "example": "le-petit-prince"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un conte poétique et philosophique sous l'apparence d'un livre pour enfants.",
                        "nullable": true
                    },
                    "total_copies": {
                        "type": "integer",
                        "example": 10
                    },
                    "available_copies": {
                        "type": "integer",
                        "example": 7
                    },
                    "degraded_copies": {
                        "type": "integer",
                        "example": 1
                    },
                    "views": {
                        "type": "integer",
                        "example": 152
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "title": "Category",
                "description": "Model for book categories",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Science Fiction"
                    },
                    "slug": {
                        "description": "URL-friendly slug of the category",
                        "type": "string",
                        "example": "science-fiction"
                    },
                    "created_at": {
                        "description": "Timestamp when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the category was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}